--- src/main/java/corecord/dev/domain/dto/request/UserRequest.java
@@ -1,4 +1,0 @@
-package corecord.dev.domain.dto.request;
-
-public class UserRequest {
-}
