--- src/main/java/corecord/dev/domain/dto/response/UserResponse.java
@@ -1,4 +1,0 @@
-package corecord.dev.domain.dto.response;
-
-public class UserResponse {
-}
