+++ src/main/java/corecord/dev/domain/user/dto/request/UserRequest.java
@@ -1,0 +1,4 @@
+package corecord.dev.domain.user.dto.request;
+
+public class UserRequest {
+}
