+++ src/main/java/corecord/dev/domain/user/dto/response/UserResponse.java
@@ -1,0 +1,4 @@
+package corecord.dev.domain.user.dto.response;
+
+public class UserResponse {
+}
