+++ src/main/java/corecord/dev/domain/user/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package corecord.dev.domain.repository;
+package corecord.dev.domain.user.repository;
 
-import corecord.dev.domain.entity.User;
+import corecord.dev.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
